Make SubprocessClientStubs reusable
