+++ Tests/SubCommandsTests/AddModule/AddModuleTrait.swift
@@ -136,7 +136,7 @@ extension Trait where Self == AddModuleTrait {
     static func addModuleEnvironmentMock(
         pathClientStub: PathStub.Configuration = .defaultTemporary,
         nooraClientStubs: AddModuleTrait.NooraClientStubs = .init(),
-        subprocessClientStubs: AddModuleTrait.SubprocessClientStubs = .init(),
+        subprocessClientStubs: SubprocessClientStubs = .init(),
         configClientStubs: ConfigFileStub = .init(),
         clientErrorStub: AddModuleTrait.ClientErrorStub? = nil
     ) -> Self {
@@ -176,79 +176,5 @@ extension AddModuleTrait {
             self.selectDependencies = selectDependencies
             self.dependencies = dependencies
         }
-    }
-
-    struct SubprocessClientStubs {
-        let packageDump: Data
-        let showDependencies: Data
-
-        init(
-            packageDump: String = SubprocessClientStubs.packageJSON,
-            showDependencies: String = SubprocessClientStubs.dependenciesGraph
-        ) {
-            self.packageDump = Data(packageDump.utf8)
-            self.showDependencies = Data(showDependencies.utf8)
-        }
-
-        // TODO: Find a better way to handle stubs for reusable client methods
-        func result(for command: ShellCommand) -> Data {
-            switch command {
-                case .swift(.package(.dumpPackage, _)):
-                    return packageDump
-                case .swift(.package(.showDependencies(.json), _)):
-                    return showDependencies
-                default:
-                    reportIssue("Invalid command stub.")
-                    return Data("{}".utf8)
-            }
-        }
-    }
-}
-
-extension AddModuleTrait.SubprocessClientStubs {
-    static var packageJSON: String {
-        """
-        {
-          "name": "StubPackage",
-          "products": [
-            {
-              "name": "ProductA",
-              "type": { "library": ["automatic"] },
-              "targets": ["TargetA"]
-            },
-            {
-              "name": "ProductB",
-              "type": { "library": ["automatic"] },
-              "targets": ["TargetB"]
-            }
-          ],
-          "targets": [
-            {
-              "name": "TargetA",
-              "type": "regular"
-            },
-            {
-              "name": "TargetB",
-              "type": "regular"
-            },
-            {
-              "name": "TargetC",
-              "type": "test"
-            }
-          ]
-        }
-        """
-    }
-
-    static var dependenciesGraph: String {
-        """
-        {
-          "dependencies": [
-            {
-              "path": "/path/to/DependencyA"
-            }
-          ]
-        }
-        """
     }
 }
